implement dotmatrix timer

+++ palette/dotmatrix.c
@@ -19,7 +19,7 @@ void init_dotmatrix(){
 	}
 }
 
-void dotmatrix_write(int count){
+void dotmatrix_write(unsigned short count){
 
     if(!(DOTMATRIX_ON>0)){ // TEXTLCD off
         // do nothing
